+++ assets/javascripts/kitten/index.js
@@ -259,8 +259,6 @@ export { TitleWithStroke } from './components/typography/title-with-stroke'
 // Uploaders
 export { BasicUploader } from './components/uploaders/basic-uploader'
 export { DocumentManager } from './components/uploaders/document-manager'
-export { SimpleUploader } from './components/uploaders/simple-uploader'
-export { Uploader } from './components/uploaders/uploader'
 
 // Videos
 export { Video } from './components/videos/video'